@@ -20,7 +20,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 2.1.0+15
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   flutter:
@@ -29,18 +29,17 @@ dependencies:
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
-  cupertino_icons: ^0.1.3
-  flutter_email_sender: ^3.0.1
-  intl:
-  path_provider:
-  shared_preferences: '>=0.5.8 <2.0.0'
-  # traindown: ^3.0.0
+  cupertino_icons: ^1.0.3
+  flutter_email_sender: ^5.0.2
+  intl: ^0.17.0
+  path_provider: ^2.0.2
+  shared_preferences: ^2.0.6
   traindown: ^3.1.0
-  url_launcher: ^5.5.0
+  url_launcher: ^6.0.9
   sqflite: ^2.0.0+3
 
 dev_dependencies:
-  flutter_launcher_icons: ^0.7.5
+  flutter_launcher_icons: ^0.9.0
   flutter_test:
     sdk: flutter
   lints: ^1.0.1
